Move implementation into len_zero.rs
